feat: migrate to oid4vc-core
