+++ oid4vc-core/Cargo.toml
@@ -1,24 +1,16 @@
 [package]
-name = "siopv2"
+name = "oid4vc-core"
 version = "0.1.0"
 edition = "2021"
-description = "Rust implementation of the OpenID Connect Self-Issued OpenID Provider v2 standard"
-homepage.workspace = true
-keywords = ["oid4vc", "openid4vc", "siopv2", "OpenID"]
-license.workspace = true
-repository.workspace = true
 
 [dependencies]
-oid4vc-core = { path = "../oid4vc-core" }
-oid4vp = { path = "../oid4vp" }
-oid4vci = { path = "../oid4vci" }
 tokio.workspace = true
 serde.workspace = true
-serde_json.workspace = true
-serde_with.workspace = true
-getset.workspace = true
+serde_json = "1.0"
+serde_with = "2.3"
 anyhow = "1.0.70"
 chrono = "0.4.24"
+getset = "0.1.2"
 jsonwebtoken = "8.2.0"
 reqwest = { version = "0.11.14", default-features = false, features = ["json", "rustls-tls"] }
 base64-url = "2.0.0"
@@ -30,11 +22,3 @@ serde_urlencoded = "0.7.1"
 derive_more = "0.99.16"
 identity_credential = { git = "https://git@github.com/iotaledger/identity.rs", rev = "4f27434" } 
 futures = "0.3"
-
-[dev-dependencies]
-ed25519-dalek = "1.0.1"
-rand = "0.7"
-lazy_static = "1.4.0"
-derivative = "2.2.0"
-wiremock = "0.5.18"
-
