+++ oid4vc-core/src/authentication/mod.rs
