+++ oid4vc-core/src/authentication/sign.rs
