+++ oid4vc-core/src/authentication/subject.rs
