+++ oid4vc-core/src/authentication/validator.rs
