+++ oid4vc-core/src/authentication/verify.rs
