+++ oid4vc-core/src/collection.rs
