+++ oid4vc-manager/tests/siopv2/implicit.rs
@@ -1,14 +1,15 @@
 use crate::common::{MemoryStorage, MockSubject, Storage};
 use chrono::{Duration, Utc};
 use lazy_static::lazy_static;
+use oid4vc_core::{DidMethod, SubjectSyntaxType};
 use oid4vc_manager::{ProviderManager, RelyingPartyManager};
 use siopv2::{
     claims::{Address, IndividualClaimRequest},
+    relying_party::ResponseItems,
     request::ResponseType,
     scope::ScopeValue,
-    subject_syntax_type::DidMethod,
     AuthorizationRequest, AuthorizationResponse, ClientMetadata, RequestUrl, Scope, StandardClaimsRequests,
-    StandardClaimsValues, SubjectSyntaxType,
+    StandardClaimsValues,
 };
 use std::{str::FromStr, sync::Arc};
 use wiremock::{
@@ -169,7 +170,7 @@ async fn test_implicit_flow() {
     // The `RelyingParty` then validates the response by decoding the header of the id_token, by fetching the public
     // key corresponding to the key identifier and finally decoding the id_token using the public key and by
     // validating the signature.
-    let (id_token, _) = relying_party_manager.validate_response(&response).await.unwrap();
+    let ResponseItems { id_token, .. } = relying_party_manager.validate_response(&response).await.unwrap();
     assert_eq!(
         id_token.standard_claims().to_owned(),
         StandardClaimsValues {
