+++ oid4vci/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod credential;
+
+pub use credential::{VerifiableCredentialJwt, VerifiableCredentialJwtBuilder};
