+++ oid4vp/src/lib.rs
@@ -1,4 +1,8 @@
+pub mod oid4vp_params;
+pub mod token;
+
 pub use dif_presentation_exchange::{
     evaluate_input, ClaimFormatDesignation, InputDescriptor, InputDescriptorMappingObject, PathNested,
     PresentationDefinition, PresentationSubmission,
 };
+pub use {oid4vp_params::Oid4vpParams, token::vp_token::VpToken};
