+++ siopv2/src/token/id_token.rs
@@ -1,7 +1,7 @@
 use super::id_token_builder::IdTokenBuilder;
 use crate::{parse_other, StandardClaimsValues};
 use getset::Getters;
-use is_empty::IsEmpty;
+use oid4vc_core::RFC7519Claims;
 use serde::{Deserialize, Serialize};
 use serde_with::skip_serializing_none;
 
@@ -38,27 +38,6 @@ pub struct SubJwk {
     pub(super) kty: String,
     pub(super) n: String,
     pub(super) e: String,
-}
-
-/// Set of IANA registered claims by the Internet Engineering Task Force (IETF) in
-/// [RFC 7519](https://tools.ietf.org/html/rfc7519#section-4.1).
-#[skip_serializing_none]
-#[derive(Serialize, Deserialize, Debug, Default, PartialEq, Clone, IsEmpty, Getters)]
-pub struct RFC7519Claims {
-    #[getset(get = "pub")]
-    pub iss: Option<String>,
-    #[getset(get = "pub")]
-    pub sub: Option<String>,
-    #[getset(get = "pub")]
-    pub aud: Option<String>,
-    #[getset(get = "pub")]
-    pub exp: Option<i64>,
-    #[getset(get = "pub")]
-    pub nbf: Option<i64>,
-    #[getset(get = "pub")]
-    pub iat: Option<i64>,
-    #[getset(get = "pub")]
-    pub jti: Option<String>,
 }
 
 #[cfg(test)]
