+++ siopv2/src/token/mod.rs
@@ -1,4 +1,2 @@
 pub mod id_token;
 pub mod id_token_builder;
-pub mod vp_token;
-pub mod vp_token_builder;
